started implementing BSM call price grid

+++ xyvals.h
@@ -75,7 +75,7 @@ struct PriceGrid
 {
 	std::string_view m_gridName{ "None" };
 	std::string_view m_xLabel{ "None" };
-	std::vector<double> m_xVals{};
+	std::vector<int> m_xVals{};
 	std::list < std::vector<double> > m_gridVals;
 };
